Added a simple test script
This will simply start the server and the client, write a random file
to the input folder. It'll show a md5sum of both the input and the output
file. Currently it'll only create a file of 1024 bytes. Might modify it
later on so it'll create a random amount of files with random file sizes.